Module grid_forge::gen::collapse::overlap

Overlapping collapsible generation

Collapsible generation algorithm working with rulesets that are based on patterns of tile type patterns. Often named within other sources as overlapping Wave-Function Collapse.

As it is based on multi-tile patterns, it is useful for generation where tiles itself don’t have a clear adjacencies relation, but instead are grouped in meaningful patterns.

Struct types

In general the types are described in the documentation for collapse module.

AdjacencyRules and FrequencyHints are self-descriptive, and both can be created by the analyzer.

Analyzer can generate AdjacencyRules, FrequencyHints and PatternCollection.

PatternCollection is a collection of OverlappingPatterns gathered from sample maps. It is an additional element over singular workflow needed for translation between collapsed patterns and underlying individual tile_type_id.

CollapsiblePatternGrid is the collection of CollapsiblePattern, and is used by Resolver to generate new map.

Structs

AdjacencyRules

Adjacency rules for the overlapping pattern-based generative algorithm.

Analyzer

GridMap analyzer for overlapping pattern data.

CollapseHistoryItem

Event in the history of tile generation process.

CollapseHistorySubscriber

Simple subscriber to collect history of tile generation process.

CollapsiblePattern

Tile which can be collapsed into one of mutliple OverlappingPattern.

CollapsiblePatternGrid

FrequencyHints

Frequency hints for the overlapping pattern-based generative algorithm.

OverlappingPattern3D

OverlappingPattern for three-dimensional grids.

OverlappingPatternGrid

Grid containing pattern data derived from original GridMap2D.

PatternCollection

Collection holding all found patterns found in sample maps.

Resolver

Enums

PatternTileData

Tile data of inner grid within OverlappingPatternGrid.

Traits

OverlappingPattern

Pattern used in Overlapping Collapse algorithm.

Subscriber

When applied to the struct allows injecting it into overlap::Resolver to react on each tile being collapsed.

Type Aliases

OverlappingPattern2D

OverlappingPattern for two-dimensional grids.