Trait grid_forge::gen::collapse::overlap::OverlappingPattern

pub trait OverlappingPattern
where
    Self: Clone + PartialEq + Eq + Hash + Debug + Sealed,
{
    const X_LEN: usize;
    const Y_LEN: usize;
    const Z_LEN: usize;

    // Required methods
    fn pattern_id(&self) -> u64;
    fn tile_type_id(&self) -> u64;
    fn get_id_for_pos(
        &self,
        anchor_pos: &GridPosition,
        pos: &GridPosition
    ) -> u64;
    fn is_compatible_with(&self, other: &Self, direction: GridDir) -> bool;
    fn secondary_tile_positions(anchor_pos: &GridPosition) -> Vec<GridPosition>;
}

Pattern used in Overlapping Collapse algorithm.

To declare the size of pattern to use, specify the dimensionality by providing OverlappingPattern2D or OverlappingPattern3D type signature in the dependent type declarations. Sizes of 2 and 3 tiles are the most efficient and provide most interesting outputs, though it all depends of the structute in the input grids.

Specifying a size of 1 in all directions will provide totally random outputs. For the single-tiled collapsible generative algorithm the singular-based should be always preferred.

It is comprised of tile_type_ids of all tiles found in the range of the pattern, where the one present at [0][0][0] position is considered the main tile, and all others are secondary tiles.

Algorithm selecting a pattern to be present in some place will result in the main tile to be placed there, while secondary tiles are there to check compatibility beetween two different OverlappingPatterns.

Required Associated Constants

const X_LEN: usize

Size of the pattern on the x axis.

const Y_LEN: usize

Size of the pattern on the y axis.

const Z_LEN: usize

Size of the pattern on the z axis.

Required Methods

fn pattern_id(&self) -> u64

Retrieves pattern identifier.

fn tile_type_id(&self) -> u64

Retrieves tile_type_id of pattern primary tile.

fn get_id_for_pos(&self, anchor_pos: &GridPosition, pos: &GridPosition) -> u64

Gets tile_type_id for a TileData of a tile present in the pattern, given the GridPosition of the primary tile (anchor_pos) and specific position (pos).

Panic

This method will panic if the pos is located beyond boundaries of the pattern.

fn is_compatible_with(&self, other: &Self, direction: GridDir) -> bool

Checks compatibility between two patterns is specified direction.

fn secondary_tile_positions(anchor_pos: &GridPosition) -> Vec<GridPosition>

Retrieves positions of the secondary tiles of the pattern.

Object Safety

This trait is not object safe.