Struct grid_forge::gen::collapse::singular::CollapsibleTileGrid

pub struct CollapsibleTileGrid<Tile: IdentifiableTileData> { /* private fields */ }

Collapsible grid compatible with singular::Resolver.

It stores the data of the tiles in the internal grid of CollapsibleTile. Holds information about the rules for the generation of the tiles and the weights of the options, provide options to populate the grid with some collapsed tiles before the generation process and retrieve the collapsed tiles after the generation ends.

Example

use grid_forge::gen::collapse::*;
use grid_forge::{GridPosition, GridSize};
use grid_forge::identifiable::*;
use grid_forge::identifiable::builders::*;

// Create new empty grid.
let mut collapsible_grid = singular::CollapsibleTileGrid::new_empty(GridSize::new_xy(10, 10), &frequency_hints, &adjacency_rules);

// We can prepopulate existing collapsible grid with some collapsed tiles using `CollapsedGrid`.
let mut collapsed_grid = CollapsedGrid::new(GridSize::new_xy(10, 10));
collapsed_grid.insert_data(&GridPosition::new_xy(0, 0), CollapsedTileData::new(0));
collapsed_grid.insert_data(&GridPosition::new_xy(1, 0), CollapsedTileData::new(1));

collapsible_grid.populate_from_collapsed(&collapsed_grid).unwrap();

// The collapsible grid can be created directly from a `CollapsedGrid`.
let mut collapsible_grid = singular::CollapsibleTileGrid::new_from_collapsed(&collapsed_grid, &frequency_hints, &adjacency_rules).unwrap();

// If there is a need to change the rules after the grid is created, it can be done using `change` method.
collapsible_grid = collapsible_grid.change(&new_frequency_hints, &adjacency_rules).unwrap();

// The grid can be retrieved as a `CollapsedGrid`.
let collapsed_grid = collapsible_grid.retrieve_collapsed();

// The grid can be retrieved as a `GridMap2D` using compatible `IdentTileBuilder`.
let ident_grid = collapsible_grid.retrieve_ident(&IdentTileTraitBuilder::<BasicIdentTileData>::default());

Implementations

impl<Tile: IdentifiableTileData> CollapsibleTileGrid<Tile>

pub fn new_empty( size: GridSize, frequencies: &FrequencyHints<Tile>, adjacencies: &AdjacencyRules<Tile> ) -> Self

Creates a new empty grid with given GridSize, preparing the rules for the generation of the tiles and the weights of the options.

pub fn new_from_collapsed( collapsed: &CollapsedGrid, frequencies: &FrequencyHints<Tile>, adjacencies: &AdjacencyRules<Tile> ) -> Result<Self, CollapsibleGridError>

Creates a new grid using the CollapsedGrid as a source grid. Created grid will have the same size as the source grid, and will be populated with existing collapsed tiles.

Method can return an error if the collapsed grid contains tiles with tile_type_ids that are not present in the provided frequency hints and adjacency rules.

pub fn change( self, frequencies: &FrequencyHints<Tile>, adjacencies: &AdjacencyRules<Tile> ) -> Result<Self, CollapsibleGridError>

Changes the rules for the generation of the tiles and the weights of the options.

Method can return an error if the inner collapsible grid contains tiles with tile_type_ids that are not present in the provided frequency hints and adjacency rules.

pub fn populate_from_collapsed( &mut self, collapsed: &CollapsedGrid ) -> Result<(), CollapsibleGridError>

Populates the grid with all collapsed tiles from the provided CollapsedGrid.

Method can return an error if the provided grid contains tiles with tile_type_ids that are not present in the provided frequency hints and adjacency rules or the provided grid size is greater than the size of inner collapsible grid.