Function grid_forge::godot::load_gridmap_from_tilemap_manual

pub fn load_gridmap_from_tilemap_manual<Data: IdentifiableTileData, B: IdentTileBuilder<Data>>(
    tilemap: &Gd<TileMap>,
    collection: &GodotTileMapCollection,
    builder: &B
) -> Result<GridMap2D<Data>, GodotTileError>

Loads GridMap2D from [TileMap], using tiles collected in GodotTileMapCollection.

Currently supports only layer = 0.

As the process is manual, the tile_type_id for each tile will be taken from the collection, not generated automatically. Process can fail if the collection does not contain the required tile.

If such control over the process is not needed, load_gridmap_from_tilemap_auto can be used.