grid_forge

Module identifiable

This module provide a way to identify a tile by its tile_type_id.

As for exact meaning of Tile Type it is mainly left to the direct implementation and use-case, eg. sometimes just the biome (like grass tile, water type, desert type etc.) will be distinct types, other times more specific identity will be needed. All these details are left for implementor to decide upon.

One thing that should never be taken into account is specific position of the tile in the grid, as it would mean every tile on every grid will be of different type.

tile_type_id is of u64 type, which makes it easily implementable by hashing some specific properties of the tile present on the struct.

Modules

builders

IdentifiableTileData instance builders.

collection

Structs

BasicIdentTileData

Basic tile struct that implements IdentifiableTileData, holding only the most basic information.

Traits

IdentifiableTileData

Its implementation makes the specific tile identifiable and discernable from other tile instances in regards to tile type. For the generative algorithms using this trait to match and select tiles, general rules of the tile identity when implementing this trait manually should be: