grid_forge::identifiable

Module builders

IdentifiableTileData instance builders.

Many methods in grid_forge needs to construct new instances of GridMap2D and fill them with new instances of tiles by their tile_id. For them to be flexible, they need to use a builder struct, using some strategy to create new tile instances.

User can create their own IdentTileBuilder-implementing struct to use their own method of building new tiles, though there are already some builders provided, using some basic strategies.

Structs

IdentTileCloneBuilder

IdentTileBuilder which creates new tiles of Clone-implementing tile struct. Prototype of tile with each tile_id need to be provided to the builder via add_tiles.

IdentTileFunBuilder

IdentTileBuilder which creates new tiles with given identifier based on the contructor functions provided to the to the builder via set_tile_constructor.

IdentTileTraitBuilder

IdentTileBuilder which creates new tiles with given identifier based on the tile implementation of ConstructableViaIdentifierTile.

TileBuilderError

Error stemming from missing tiles in IdentTileBuilder.

Traits

ConstructableViaIdentifierTile

Trait which allows creating new istance of struct implementing IdentifiableTileData.

IdentTileBuilder

Trait shared by objects which on basis of the grid position and tile identifier of given IdentifiableTileData-implementing struct can create correct instance of the tile. Necessary for many GridMap2D creating methods.