Trait grid_forge::identifiable::builders::ConstructableViaIdentifierTile

pub trait ConstructableViaIdentifierTile
where
    Self: IdentifiableTileData,
{
    // Required method
    fn tile_new(tile_type_id: u64) -> Self;
}

Trait which allows creating new istance of struct implementing IdentifiableTileData.

See also BasicIdentTileData for basic identifiable tile type which implements this trait out of the box whenever you don’t need your tile struct to hold any additional information.

Examples

Implementing this trait for you custom tile makes it constructable via IdentTileTraitBuilder.

use grid_forge::{GridPosition, TileData, TileContainer};
use grid_forge::identifiable::IdentifiableTileData;
use grid_forge::identifiable::builders::{ConstructableViaIdentifierTile, IdentTileBuilder, IdentTileTraitBuilder};

// Implementing custom TileData supported by `IdentTileCloneBuilder`.
#[derive(Clone)]
struct MyTileData {
    tile_type_id: u64,
}

impl TileData for MyTileData {};

impl IdentifiableTileData for MyTileData {
    fn tile_type_id(&self) -> u64 {
        self.tile_type_id
    }
}

impl ConstructableViaIdentifierTile for MyTileData {
    fn tile_new(tile_type_id: u64) -> Self {
        Self { tile_type_id }
    }
}

let builder = IdentTileTraitBuilder::<MyTileData>::default();
let tile = builder.build_tile_unchecked(GridPosition::new_xy(2,3), 45);

assert_eq!(GridPosition::new_xy(2,3), tile.grid_position());
assert_eq!(45, tile.as_ref().tile_type_id());

Required Methods

fn tile_new(tile_type_id: u64) -> Self

Object Safety

This trait is not object safe.