Trait grid_forge::identifiable::builders::IdentTileBuilder

pub trait IdentTileBuilder<Data: IdentifiableTileData> {
    // Required methods
    fn build_tile_unchecked(
        &self,
        position: GridPosition,
        tile_type_id: u64
    ) -> GridTile<Data>;
    fn build_tile(
        &self,
        position: GridPosition,
        tile_type_id: u64
    ) -> Result<GridTile<Data>, TileBuilderError>;
    fn check_missing_ids(
        &self,
        tile_type_ids: &[u64]
    ) -> Result<(), TileBuilderError>;
}

Trait shared by objects which on basis of the grid position and tile identifier of given IdentifiableTileData-implementing struct can create correct instance of the tile. Necessary for many GridMap2D creating methods.

Three different builders are available in the grid_forge:

IdentTileFunBuilder - for tiles not implementing any additional traits.

IdentTileCloneBuilder - for tiles implementing Clone.

IdentTileTraitBuilder - for tiles implementing ConstructableViaIdentifierTile.

The logic for building tile is encapsulated in build_tile_unchecked and build_tile methods. The unchecked version is usually faster and is recommended to be used when creating a batch of tiles at once - can be used reliably if check_missing_ids is called before beginning the batch operation.

Required Methods

fn build_tile_unchecked( &self, position: GridPosition, tile_type_id: u64 ) -> GridTile<Data>

Creates tile with given tile identifier at given grid position.

Panics

Can panic if builder does not have possibility to construct tile of given tile_id based on the gathered information. You can check for missing tile ids with check_missing_ids or use its fallible version: build_tile.

fn build_tile( &self, position: GridPosition, tile_type_id: u64 ) -> Result<GridTile<Data>, TileBuilderError>

Creates tile with given tile identifier at given grid position. Returns error if cannot construct tile of given tile_id.

fn check_missing_ids( &self, tile_type_ids: &[u64] ) -> Result<(), TileBuilderError>

Checks for missing tile creators out of provided slice of tile_id.