Trait grid_forge::identifiable::collection::IdentTileCollection

pub trait IdentTileCollection {
    type DATA: Hash;

    // Required methods
    fn inner(&self) -> &HashMap<u64, Self::DATA>;
    fn inner_mut(&mut self) -> &mut HashMap<u64, Self::DATA>;
    fn rev(&self) -> &HashMap<u64, u64>;
    fn rev_mut(&mut self) -> &mut HashMap<u64, u64>;

    // Provided methods
    fn on_add(&mut self, _data: &Self::DATA) { ... }
    fn on_remove(&mut self, _data: &Self::DATA) { ... }
    fn add_tile(&mut self, data: Self::DATA) -> bool { ... }
    fn set_tile(&mut self, data: Self::DATA) -> bool { ... }
    fn remove_tile(&mut self, data: &Self::DATA) -> bool { ... }
    fn add_tile_data(&mut self, tile_type_id: u64, data: Self::DATA) -> bool { ... }
    fn set_tile_data(
        &mut self,
        tile_type_id: u64,
        data: Self::DATA
    ) -> Option<Self::DATA> { ... }
    fn remove_tile_data(&mut self, tile_type_id: &u64) -> Option<Self::DATA> { ... }
    fn get_tile_data(&self, tile_type_id: &u64) -> Option<&Self::DATA> { ... }
    fn get_tile_type_id(&self, data: &Self::DATA) -> Option<u64> { ... }
    fn generate_type_id(data: &Self::DATA) -> u64 { ... }

}

Trait implementing the behaviour for easily keeping and retrieving the tile_type_id alongside some arbitrary DATA kept for the tile.

Objects implementing this trait are often used for operations of transferring between non-grid-forge representation of grid map and
GridMap2D alongside builders implementing IdentTileBuilder trait.

For the collection to work properly, both tile_type_id needs to be unique for tile of each type (as required by identifiable module assumptions), as well as the DATA itself.

Automatic derivement of tile_type_id

As the external representations rarely hold a specific tile_type_id, the collection handles storing the DATA without specifying the identifier manually (see: add_tile and set_tile). It calculates the type_tile_id by hashing the underlying data using DefaultHasher.

This makes the collection instantly useful, but only during transfers between one external representation and the GridMap2D, as synchronizing the tile_type_id between multiple collections is very difficult.

Manual specification of tile_type_id

To synchronize between multiple external representations of grid map, with GridMap2D working as an intermediate step, the DATA should be stored in controlled way in multiple IdentTileCollections to keep the tile_type_id synchronized (see: add_tile_data and set_tile_data).

Prefer specific facade methods

Methods of this trait are intended to be generic and low-level. Specific implementors can wrap them below additional logic. If these kind of methods are available, prefer to make use of them.

Required Associated Types

type DATA: Hash

External data kept within the collection, bind to the specific tile_type_id.

Required Methods

fn inner(&self) -> &HashMap<u64, Self::DATA>

Exposes inner hashmap. Necessary for other methods implementations.

fn inner_mut(&mut self) -> &mut HashMap<u64, Self::DATA>

Exposes inner hashmap mutably. Necessary for other methods implementations.

fn rev(&self) -> &HashMap<u64, u64>

Exposes reverse hashmap. Necessary for other methods implementations.

fn rev_mut(&mut self) -> &mut HashMap<u64, u64>

Exposes reverse hashmap mutably. Necessary for other methods implementations.

Provided Methods

fn on_add(&mut self, _data: &Self::DATA)

Additional operation that will be done during addition of new DATA into the collection. No-op by default.

fn on_remove(&mut self, _data: &Self::DATA)

Additional operation that will be done during removal of old DATA from the collection. No-op by default.

fn add_tile(&mut self, data: Self::DATA) -> bool

Adds tile data without tile_type_id provided - it will be generated with generate_type_id. If either data or the generated tile_type_id are already present in the collection, addition will be skipped, returning false.

fn set_tile(&mut self, data: Self::DATA) -> bool

Sets tile data without tile_type_id provided - it will be generated with generate_type_id. If either data or generated tile_type_id are already present in the collection, they will be overwritten, returning true.

fn remove_tile(&mut self, data: &Self::DATA) -> bool

Removes tile data if either provided data or hash generated with generate_type_id is present in the collection.

fn add_tile_data(&mut self, tile_type_id: u64, data: Self::DATA) -> bool

Adds data for specified tile_type_id. If given tile_type_id is already present, no changes are made and returns false;

fn set_tile_data( &mut self, tile_type_id: u64, data: Self::DATA ) -> Option<Self::DATA>

Sets data for specified tile_type_id. Returns removed data stored for specified id, if present.

fn remove_tile_data(&mut self, tile_type_id: &u64) -> Option<Self::DATA>

Removes DATA for provided tile_type_id. Returns removed data.

fn get_tile_data(&self, tile_type_id: &u64) -> Option<&Self::DATA>

Gets DATA stored for given tile_type_id.

fn get_tile_type_id(&self, data: &Self::DATA) -> Option<u64>

Gets tile_type_id bind to given data.

fn generate_type_id(data: &Self::DATA) -> u64

Generates tile_type_id using provided DATA.

Object Safety

This trait is not object safe.