Trait grid_forge::TileContainer

pub trait TileContainer {
    // Required method
    fn grid_position(&self) -> GridPosition;
}

Trait gathering the containers for TileData outside of the GridMap2D.

Allows accessing all the data not contained within tile data, making sense only in context of the grid map.

Required Methods

fn grid_position(&self) -> GridPosition