Module grid_forge::vis::ops

Various IO operations transforming between GridMap2D and ImageBuffer representation of grid map.

Functions

check_grid_image_size

Checks the size of the ImageBuffer before writing GridMap2D visual representation into it. Results in VisError if the image buffer size and GridSize does not match.

check_grid_vis_size

Checks the size of the ImageBuffer while loading GridMap2D from its visual representation, and produces the GridSize inferred from the image size. Results in VisError if the image size is not compatible with provided tile size in pixels.

init_map_image_buffer

Utility function to generate ImageBuffer of correct size for specific size of GridMap2D to write into with write_gridmap_identifiable and write_gridmap_vis.

load_gridmap_identifiable_auto

Easily load GridMap2D of IdentifiableTileData-implementing TileData, automatically saving each tile into provided VisCollection.

load_gridmap_identifiable_manual

Load GridMap2D of IdentifiableTileData-implementing struct using registered pixels in VisCollection.

write_gridmap_identifiable

Write GridMap2D comprised of tiles containing IdentifiableTileData into provided ImageBuffer, using pixel data gatheed in VisCollection.

write_gridmap_vis

Write GridMap2D comprised of tiles containing VisTileData into provided ImageBuffer. Pixel data retrieved via VisTileData::vis_pixels will be used.