grid_forge::vis::ops

Function load_gridmap_identifiable_auto

pub fn load_gridmap_identifiable_auto<Data, P, B, const WIDTH: usize, const HEIGHT: usize>(
    image_buffer: &ImageBuffer<P, Vec<P::Subpixel>>,
    collection: &mut VisCollection<P, WIDTH, HEIGHT>,
    builder: &B,
) -> Result<GridMap2D<Data>, VisError<WIDTH, HEIGHT>>
where
    Data: IdentifiableTileData,
    P: PixelWithDefault + 'static,
    B: IdentTileBuilder<Data>,

Easily load GridMap2D of IdentifiableTileData-implementing TileData, automatically saving each tile into provided VisCollection.

Arguments

image_buffer - an ImageBuffer containing the source image data.

collection - a VisCollection containing the loaded tile pixels. New tile pixels will be automatically inserted, calculating the tile_type_id on basis of tile pixels.

builder - a struct which can be used to construct new tiles on basis of their tile_id. One of IdentTileBuilder implementing objects.

Warning

As the tile_type_id is automatically calculated with this function on basis of pixels, it won’t work with specific, manually declared identifiers. In this case, you need to use load_gridmap_identifiable_manual.