Function grid_forge::vis::ops::load_gridmap_identifiable_manual

pub fn load_gridmap_identifiable_manual<Data, P, B, const WIDTH: usize, const HEIGHT: usize>(
    image_buffer: &ImageBuffer<P, Vec<P::Subpixel>>,
    collection: &VisCollection<P, WIDTH, HEIGHT>,
    builder: &B
) -> Result<GridMap2D<Data>, VisError<WIDTH, HEIGHT>>
where
    Data: IdentifiableTileData,
    P: PixelWithDefault + 'static,
    B: IdentTileBuilder<Data>,

Load GridMap2D of IdentifiableTileData-implementing struct using registered pixels in VisCollection.

If you don’t need to declare the tile_type_id yourself, you can use load_gridmap_identifiable_auto, which generates identifiers on basis of each pixel.

Arguments

image_buffer - an ImageBuffer containing the source image data.

collection - a VisCollection containing the tile pixels. It needs to be populated before usage.

builder - a struct which can be used to construct new tiles on basis of their tile_id. One of IdentTileBuilder implementing objects.