grid_forge::gen

Module collapse

Collapsible procedural generation

Algorithms contained within this module works by having a collection of tiles, which final identity is unknown in the beginning - every position can be any tile. In the process of generating a new grid, algorithm will collapse possible options for position, choosing randomly one of them. Collapsed tile will then put contraints over the remaining tiles, reducing the number of possible options. Process will then continue until all tiles are collapsed into one possible option.

This process is called elsewhere most often Model Synthesis or Wave Function Collapse, though the distinction between the two is not always clear. Implementation contained there can be tailored to specific needs for the generation process, which makes the line even more blurry, so new name is used here.

Main distinction

Two main types of algorithms are provided in distinct submodules:

singular - implementation that can be used to generate maps. Each tile type needs to be self-descriptive in regards to its possible neighbours. Constraints for this type of generation is based strictly on possible neighbours for each given tile, set of rules which can be described as: tile X can be placed in direction D of tile Y.

overlap - implementation that can be used to generate maps based on observed patterns. Direct possible adjacents of each tile are less important there, the pattern context is more important. First step for this type of generation is to create a collection of patterns from sample gridmaps, which then will be tested for possible neigbours. During the process the possible patterns will then be collapsed instead of individual tiles.

Struct types

Structs contained in submodules mentioned above are categorized in analogous way, with some additional distinctions described further in documentation for concrete struct.

adjacency rules are the rules that are used to determine possible neighbours for each tile.

frequency hints can be described as a weights per option - they can be derived from sample gridmaps and are a way to influence the frequency of options occuring in the generated grid.

analyzers provide methods for analyzing sample gridmaps and creating rulesets out of them.

collapsible grids are the source of information for the resolvers, from the collection of collapsible tiles to the prepared adjacency rules and frequency hints. They can also contain some pre-collapsed tiles, providing initial constraints for the generation.

resolvers are the main executors of the algorithm and are responsible for collapsing the tiles in the collapsible grids.

queues are used to determine the order in which tiles are collapsed: PositionQueue takes next position to collapse in a fixed order, while EntrophyQueue fetch the next position to collapse with the lowest entrophy.

Modules

overlap

Overlapping collapsible generation

singular

Singular collapsible generation

Structs

CollapseError

Error occuring during collapse process.

CollapsedGrid

GridMap2D containing data of CollapsedTileData.

CollapsedTileData

Simple TileData containing only the tile_type_id.

DebugSubscriber

Basic Subscriber for debugging purposes.

EntrophyQueue

Select next position to collapse using smallest entrophy condition.

PositionQueue

A queue that collapses tiles consecutively in a fixed direction, based solely on their position.

Enums

PositionQueueDirection

Enum defining the direction in which the tiles will be collapsed.

PositionQueueStartingPoint

Enum defining the starting point of the collapse wave.

Traits

CollapseQueue

Trait shared by objects that handle the selecting algorithm for next tile to collapse within collapse resolvers.

CollapsibleGrid

Trait shared by a structs holding a grid of CollapsibleTileData, useable by dedicated resolvers to collapse the grid.

CollapsibleTileData

Trait shared by TileData used within collapsible generative algorithms.