grid_forge::gen::collapse

Module singular

Singular collapsible generation

Collapsible generation algorithm working with rulesets that are based on tile type adjacencies. Often named within other sources as single-tiled algorithms, both Model Synthesis and Wave-Function Collapse can fall into this category.

As it is based on tile type adjacencies, it is more viable to describe the rules manually than pattern-based overlap.

Struct types

In general the types are described in the documentation for collapse module.

AdjacencyRules and FrequencyHints are self-descriptive. The latter are not produced by the analyzer, but the method for their derivation from the sample gridmap is exposed..

Analyzer is a trait implemented by two distincts analyzers. The IdentityAnalyzer in general produced more restrictive rules, as it search for exact neigbours on the sample gridmap. The BorderAnalyzer is more liberal, as it takes an extra step and derives more rules based on the distinct tile borders, making additional options available if they could be placed on the sample gridmap next to each other.

CollapsibleTileGrid is the collection of CollapsibleTile.

Resolver is the main executor of the algorithm.

Structs

AdjacencyRules

Adjacency rules for singular collapse algorithm.

BorderAnalyzer

Analyzer creating adjacency rules based on the borders between tiles.

CollapseHistoryItem

Event in the history of tile generation process, containing the GridPosition of the tile alongside its collapsed tile_type_id.

CollapseHistorySubscriber

Simple subscriber to collect history of tile generation process.

CollapsibleTile

Tile with options that can be collapsed into one of them. Mostly used within the CollapsibleTileGrid.

CollapsibleTileGrid

Collapsible grid compatible with singular::Resolver.

FrequencyHints

Frequency hints for the adjacency-based generative algorithm.

IdentityAnalyzer

Analyzer creating exact adjacency rules on basis of sample map.

Resolver

Resolver of the singular collapsible procedural algorithm.

Traits

Analyzer

Trait shared by analyzers producing AdjacencyRules.

Subscriber

When applied to the struct allows injecting it into singular::Resolver to react on each tile being collapsed.