grid_forge::gen::collapse

Trait CollapseQueue

pub trait CollapseQueue
where
    Self: Default + Sized + Sealed,
{
    // Required methods
    fn get_next_position(&mut self) -> Option<GridPosition>;
    fn initialize_queue<T: CollapsibleTileData>(
        &mut self,
        tiles: &[GridTile<T>],
    );
    fn update_queue<Tile, Data>(&mut self, tile: &Tile)
       where Tile: TileContainer + AsRef<Data>,
             Data: CollapsibleTileData;
    fn len(&self) -> usize;
    fn is_empty(&self) -> bool;
}

Trait shared by objects that handle the selecting algorithm for next tile to collapse within collapse resolvers.

Required Methods

fn get_next_position(&mut self) -> Option<GridPosition>

Pop next position for collapsing.

fn initialize_queue<T: CollapsibleTileData>(&mut self, tiles: &[GridTile<T>])

Initialize the queue based on provided tiles.

fn update_queue<Tile, Data>(&mut self, tile: &Tile)
where Tile: TileContainer + AsRef<Data>, Data: CollapsibleTileData,

Update internal based on provided tile.

fn len(&self) -> usize

Checks the current size of the inner queue.

fn is_empty(&self) -> bool

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.