grid_forge::gen::collapse

Trait CollapsibleGrid

pub trait CollapsibleGrid<IT: IdentifiableTileData, CT: CollapsibleTileData>: Sized + Sealed<CT> {
    // Required methods
    fn retrieve_collapsed(&self) -> CollapsedGrid;
    fn retrieve_ident<OT: IdentifiableTileData, B: IdentTileBuilder<OT>>(
        &self,
        builder: &B,
    ) -> Result<GridMap2D<OT>, CollapsibleGridError>;

    // Provided methods
    fn empty_positions(&self) -> Vec<GridPosition> { ... }
    fn retrieve_positions(&self, collapsed: bool) -> Vec<GridPosition> { ... }
    fn remove_uncollapsed(&mut self) { ... }
}

Trait shared by a structs holding a grid of CollapsibleTileData, useable by dedicated resolvers to collapse the grid.

Required Methods

fn retrieve_collapsed(&self) -> CollapsedGrid

Retrieves the collapsed tiles in internal grid as a CollapsedGrid.

fn retrieve_ident<OT: IdentifiableTileData, B: IdentTileBuilder<OT>>( &self, builder: &B, ) -> Result<GridMap2D<OT>, CollapsibleGridError>

Retrieves the collapsed tiles in internal grid as a GridMap2D of IdentifiableTileData.

Provided Methods

fn empty_positions(&self) -> Vec<GridPosition>

Returns all empty positions in the internal grid.

fn retrieve_positions(&self, collapsed: bool) -> Vec<GridPosition>

Returns all possitions in the internal grid holds collapsed or uncollapsed tiles are either collapsed.

fn remove_uncollapsed(&mut self)

Removes all uncollapsed tiles from the internal grid.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.